+++ src/utils/utils.js
@@ -308,21 +308,20 @@ export function sendNetworkChangeMsg(netConfig) {
  */
 export function getShowTime(time) {
   try {
-    const lang = navigator.language || navigator.languages[0];
-    let date = new Date(time);
-    let timeDate = date.toLocaleString(lang, {
-      // timeZone: 'Europe/Moscow',
-      hourCycle: "h23",
-      year: "numeric",
-      month: "2-digit",
-      day: "2-digit",
-      hour: "2-digit",
-      minute: "2-digit",
-      second: "2-digit",
-    });
-    return timeDate.replaceAll("/", "-");
+    if (!time) {
+      return "date-time";
+    }
+    const date = new Date(time);
+
+    const year = date.getFullYear();
+    const month = String(date.getMonth() + 1).padStart(2, "0"); // Months are 0-based
+    const day = String(date.getDate()).padStart(2, "0");
+    const hour = String(date.getHours()).padStart(2, "0");
+    const minute = String(date.getMinutes()).padStart(2, "0");
+
+    return `${year}-${month}-${day} ${hour}:${minute}`;
   } catch (error) {
-    return time;
+    return time
   }
 }